Refactor invoice functions and fix issues with cockroachdb tables.
